requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
colorama==0.4.6
dnspython==2.6.1
exceptiongroup==1.2.1
frozenlist==1.4.1
idna==3.7
iniconfig==2.0.0
isort==5.13.2
magic-filter==1.0.12
motor==3.4.0
multidict==6.0.5
packaging==24.0
pluggy==1.5.0
pydantic==2.5.3
pydantic_core==2.14.6
pymongo==4.6.3
pytest==8.2.0
pytest-asyncio==0.23.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
tomli==2.0.1
typing_extensions==4.11.0
yarl==1.9.4
